Main and Pause Menu
Main and Pause menu are both fully functional. Not entirely sure about the resume button since we dont have any activity in the game yet but it should work just fine. Main menu is ugly right now but it will be changed when we discuss what to make of it.